back-django/requirements.txt
absl-py==1.4.0
args==0.1.0
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
clint==0.5.1
colorama==0.4.6
comm==0.1.2
coverage==7.2.2
cryptography==39.0.2
Cython==0.29.32
debugpy==1.6.6
decorator==5.1.1
Django==3.2.18
django-cors-headers==3.14.0
djangorestframework==3.14.0
executing==1.2.0
filelock==3.9.1
flatbuffers==23.3.3
fst-pso==1.8.1
FuzzyTM==2.0.5
gast==0.4.0
gensim==4.3.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
huggingface-hub==0.13.2
idna==3.4
importlib-metadata==6.1.0
ipykernel==6.21.3
ipython==8.11.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.3.0
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==15.0.6.1
mamba==0.11.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
miniful==0.0.6
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.21
pyFUME==0.2.25
Pygments==2.14.0
pyOpenSSL==23.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
# pywin32==305
# pywinpty==2.0.10
PyYAML==6.0
# pyzmq==25.0.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
simpful==2.10.0
six==1.16.0
smart-open==6.3.0
sqlparse==0.4.3
stack-data==0.6.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
tokenizers==0.13.2
# torch==1.13.1
# torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
